+++ CHANGELOG.md
@@ -815,3 +815,4 @@
 * Move `One` module from `Algebraic` to `Operation`
 * Move `Zero` module from `Algebraic` to `Operation`
 * Move `Exponential` module from `Algebraic` to `Operation`
+* Move `<:*:>`, `<:*:<`, `>:*:>`, `>:*:<` to `Algebraic` module